requirements.txt
Flask
flask-bcrypt
httpx
perplexityai